requirements.txt
aiogram==3.4.1
Flask==3.0.0
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
aiosqlite==0.19.0
werkzeug==3.0.1
Jinja2==3.1.2aiogram==3.4.1
Flask==3.0.0
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
aiosqlite==0.19.0
werkzeug==3.0.1
Jinja2==3.1.2
